collection detail authorization

+++ collections_app/controllers/collection_detail.py
@@ -8,15 +8,26 @@ from ..models import Collection
 
 class CollectionDetail(APIView):
     """
-    Class handles the saving or updating of Collection resources.
+    Class handles the creating, saving, updating and deleting of Collection resources.
     """
 
-    def get(self, request, collectionId, format=None):
+    def get(self, request, collectionId, format=None): 
+        """
+        Returns an individual collection
+        """
+        user = request.user
         collection = Collection.objects.get(id=collectionId)
-        serializer = CollectionSerializer(collection, many=True)
+
+        if collection.user.id != user.id and collection.privacyLevel == "Private":
+            return Response({'Bad Request': 'Unauthorized. Permission Denied'}, status=status.HTTP_403_FORBIDDEN)
+
+        serializer = CollectionSerializer(collection)
         return Response(serializer.data)
 
     def post(self, request, format=None):
+        """
+        Auth users only
+        """
         user = request.user
         serializer = CollectionSerializer(data=request.data)
 
@@ -28,8 +39,12 @@ class CollectionDetail(APIView):
         return Response({'Bad Request': 'Invalid data...'}, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, collectionId):
-        # TODO Make sure only the user who owns collection is editing it
+        user = request.user
         collection = Collection.objects.get(id=collectionId)
+
+        if collection.user.id != user.id:
+            return Response({'Bad Request': 'Unauthorized. Permission Denied'}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = CollectionSerializer(collection, data=request.data)
         if serializer.is_valid():
             serializer.save(user=request.user)
@@ -38,5 +53,10 @@ class CollectionDetail(APIView):
 
     def delete(self, request, collectionId):
         collection = Collection.objects.get(id=collectionId)
+        user = request.user
+
+        if collection.user.id != user.id:
+            return Response({'Bad Request': 'Unauthorized. Permission Denied'}, status=status.HTTP_403_FORBIDDEN)
+
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
